requirements.txt
Flask==3.0.0
Flask-MySQLdb==2.0.0
qrcode[pil]==7.4.2
gunicorn==21.2.0